refactor: OrderService 구조 리팩토링(#354)
